requirements.txt
aiohttp==3.11.9
fastapi==0.115.11
openpyxl==3.1.5
Pillow==11.1.0
pydantic==2.10.6
qrcode==8.0
SQLAlchemy==2.0.36
sqlmodel==0.0.24
uvicorn==0.34.0
